+++ packages/react-querybuilder/src/components/QueryBuilder.test.tsx
@@ -55,17 +55,17 @@ export const stripQueryIds = (query: RuleGroupTypeAny): RuleGroupTypeAny =>
 const { consoleError } = consoleMocks();
 
 describe('when rendered', () => {
-  it('should have the correct role', () => {
+  it('has the correct role', () => {
     render(<QueryBuilder />);
     expect(screen.getByRole('form')).toBeInTheDocument();
   });
 
-  it('should have the correct className', () => {
+  it('has the correct className', () => {
     render(<QueryBuilder />);
     expect(screen.getByRole('form')).toHaveClass(sc.queryBuilder);
   });
 
-  it('should render the root RuleGroup', () => {
+  it('renders the root RuleGroup', () => {
     render(<QueryBuilder />);
     expect(screen.getByTestId(TestID.ruleGroup)).toBeInTheDocument();
   });
@@ -91,7 +91,7 @@ describe('when rendered with defaultQuery only', () => {
 });
 
 describe('when rendered with onQueryChange callback', () => {
-  it('should call onQueryChange with query', () => {
+  it('calls onQueryChange with query', () => {
     const onQueryChange = jest.fn<never, [RuleGroupType]>();
     const idGenerator = () => 'id';
     render(<QueryBuilder onQueryChange={onQueryChange} idGenerator={idGenerator} />);
@@ -107,7 +107,7 @@ describe('when rendered with onQueryChange callback', () => {
 });
 
 describe('when initial query without fields is provided, create rule should work', () => {
-  it('should be able to create rule on add rule click', async () => {
+  it('is able to create rule on add rule click', async () => {
     render(<QueryBuilder />);
     await user.click(screen.getByTestId(TestID.addRule));
     expect(screen.getByTestId(TestID.rule)).toBeInTheDocument();
@@ -244,7 +244,7 @@ describe('when initial query, without ID, is provided', () => {
     selectors: render(<QueryBuilder query={queryWithoutID} fields={fields} />),
   });
 
-  it('should contain a <Rule /> with the correct props', () => {
+  it('contains a <Rule /> with the correct props', () => {
     const { selectors } = setup();
     expect(selectors.getByTestId(TestID.rule)).toBeInTheDocument();
     expect(selectors.getByTestId(TestID.fields)).toHaveValue('firstName');
@@ -252,22 +252,22 @@ describe('when initial query, without ID, is provided', () => {
     expect(selectors.getByTestId(TestID.valueEditor)).toHaveValue('Test without ID');
   });
 
-  it('should have a select control with the provided fields', () => {
+  it('has a select control with the provided fields', () => {
     const { selectors } = setup();
     expect(selectors.getByTestId(TestID.fields).querySelectorAll('option')).toHaveLength(3);
   });
 
-  it('should have a field selector with the correct field', () => {
+  it('has a field selector with the correct field', () => {
     const { selectors } = setup();
     expect(selectors.getByTestId(TestID.fields)).toHaveValue('firstName');
   });
 
-  it('should have an operator selector with the correct operator', () => {
+  it('has an operator selector with the correct operator', () => {
     const { selectors } = setup();
     expect(selectors.getByTestId(TestID.operators)).toHaveValue('=');
   });
 
-  it('should have an input control with the correct value', () => {
+  it('has an input control with the correct value', () => {
     const { selectors } = setup();
     expect(selectors.getByTestId(TestID.rule).querySelector('input')).toHaveValue(
       'Test without ID'
@@ -351,12 +351,12 @@ describe('when initial operators are provided', () => {
     selectors: render(<QueryBuilder operators={operators} fields={fields} query={query} />),
   });
 
-  it('should use the given operators', () => {
+  it('uses the given operators', () => {
     const { selectors } = setup();
     expect(selectors.getByTestId(TestID.operators).querySelectorAll('option')).toHaveLength(4);
   });
 
-  it('should match the label of the first operator', () => {
+  it('matches the label of the first operator', () => {
     const { selectors } = setup();
     expect(selectors.getByTestId(TestID.operators).querySelectorAll('option')[0]).toHaveTextContent(
       'Custom Is Null'
@@ -382,20 +382,20 @@ describe('get* callbacks', () => {
   };
 
   describe('when getOperators fn prop is provided', () => {
-    it('should invoke custom getOperators function', () => {
+    it('invokes custom getOperators function', () => {
       const getOperators = jest.fn(() => [{ name: 'op1', label: 'Operator 1' }]);
       render(<QueryBuilder query={query} fields={fields} getOperators={getOperators} />);
       expect(getOperators).toHaveBeenCalledWith(rule.field, { fieldData: fields[1] });
     });
 
-    it('should handle invalid getOperators return value', () => {
+    it('handles invalid getOperators return value', () => {
       render(<QueryBuilder query={query} fields={fields} getOperators={() => null} />);
       expect(screen.getByTestId(TestID.operators)).toHaveValue('=');
     });
   });
 
   describe('when getValueEditorType fn prop is provided', () => {
-    it('should invoke custom getValueEditorType function', () => {
+    it('invokes custom getValueEditorType function', () => {
       const getValueEditorType = jest.fn(() => 'text' as const);
       render(
         <QueryBuilder query={query} fields={fields} getValueEditorType={getValueEditorType} />
@@ -405,14 +405,22 @@ describe('get* callbacks', () => {
       });
     });
 
-    it('should handle invalid getValueEditorType function', () => {
+    it('handles invalid getValueEditorType function', () => {
       render(<QueryBuilder query={query} fields={fields} getValueEditorType={() => null} />);
       expect(screen.getByTestId(TestID.valueEditor)).toHaveAttribute('type', 'text');
+    });
+
+    it('prefers valueEditorType field property', () => {
+      const checkboxFields = fields.map(f => ({ ...f, valueEditorType: 'checkbox' }));
+      render(
+        <QueryBuilder query={query} fields={checkboxFields} getValueEditorType={() => 'text'} />
+      );
+      expect(screen.getByTestId(TestID.valueEditor)).toHaveAttribute('type', 'checkbox');
     });
   });
 
   describe('when getInputType fn prop is provided', () => {
-    it('should invoke custom getInputType function', () => {
+    it('invokes custom getInputType function', () => {
       const getInputType = jest.fn(() => 'text' as const);
       render(<QueryBuilder query={query} fields={fields} getInputType={getInputType} />);
       expect(getInputType).toHaveBeenCalledWith(rule.field, rule.operator, {
@@ -420,7 +428,7 @@ describe('get* callbacks', () => {
       });
     });
 
-    it('should handle invalid getInputType function', () => {
+    it('handles invalid getInputType function', () => {
       render(<QueryBuilder query={query} fields={fields} getInputType={() => null} />);
       expect(screen.getByTestId(TestID.valueEditor)).toHaveAttribute('type', 'text');
     });
@@ -429,7 +437,7 @@ describe('get* callbacks', () => {
   describe('when getValues fn prop is provided', () => {
     const getValueEditorType = () => 'select' as const;
 
-    it('should invoke custom getValues function', () => {
+    it('invokes custom getValues function', () => {
       const getValues = jest.fn(() => [{ name: 'test', label: 'Test' }]);
       render(
         <QueryBuilder
@@ -442,7 +450,7 @@ describe('get* callbacks', () => {
       expect(getValues).toHaveBeenCalledWith(rule.field, rule.operator, { fieldData: fields[1] });
     });
 
-    it('should invoke custom getValues function returning value options', () => {
+    it('invokes custom getValues function returning value options', () => {
       const getValues = jest.fn(() => [{ value: 'test', label: 'Test' }]);
       render(
         <QueryBuilder
@@ -455,7 +463,7 @@ describe('get* callbacks', () => {
       expect(getValues).toHaveBeenCalledWith(rule.field, rule.operator, { fieldData: fields[1] });
     });
 
-    it('should generate the correct number of options', () => {
+    it('generates the correct number of options', () => {
       const getValues = jest.fn(() => [{ name: 'test', label: 'Test' }]);
       render(
         <QueryBuilder
@@ -469,7 +477,7 @@ describe('get* callbacks', () => {
       expect(opts).toHaveLength(1);
     });
 
-    it('should handle invalid getValues function', () => {
+    it('handles invalid getValues function', () => {
       // @ts-expect-error getValues should return an array of options or option groups
       render(<QueryBuilder query={query} fields={fields} getValues={() => null} />);
       const select = screen.getByTestId(TestID.valueEditor);
@@ -493,7 +501,7 @@ describe('actions', () => {
     };
   };
 
-  it('should create a new rule and remove that rule', async () => {
+  it('creates a new rule and remove that rule', async () => {
     const { selectors, onQueryChange } = setup();
     await user.click(selectors.getByTestId(TestID.addRule));
 
@@ -507,7 +515,7 @@ describe('actions', () => {
     expect(onQueryChange.mock.calls[2][0].rules).toHaveLength(0);
   });
 
-  it('should create a new group and remove that group', async () => {
+  it('creates a new group and remove that group', async () => {
     const { selectors, onQueryChange } = setup();
     await user.click(selectors.getByTestId(TestID.addGroup));
 
@@ -522,7 +530,7 @@ describe('actions', () => {
     expect(onQueryChange.mock.calls[2][0].rules).toHaveLength(0);
   });
 
-  it('should create a new rule and change the fields', async () => {
+  it('creates a new rule and change the fields', async () => {
     const { selectors, onQueryChange } = setup();
     await user.click(selectors.getByTestId(TestID.addRule));
 
@@ -533,7 +541,7 @@ describe('actions', () => {
     expect((onQueryChange.mock.calls[2][0].rules[0] as RuleType).field).toBe('field2');
   });
 
-  it('should create a new rule and change the operator', async () => {
+  it('creates a new rule and change the operator', async () => {
     const { selectors, onQueryChange } = setup();
     await user.click(selectors.getByTestId(TestID.addRule));
 
@@ -544,7 +552,7 @@ describe('actions', () => {
     expect((onQueryChange.mock.calls[2][0].rules[0] as RuleType).operator).toBe('!=');
   });
 
-  it('should change the combinator of the root group', async () => {
+  it('changes the combinator of the root group', async () => {
     const { selectors, onQueryChange } = setup();
     expect(onQueryChange.mock.calls[0][0].rules).toHaveLength(0);
 
@@ -554,7 +562,7 @@ describe('actions', () => {
     expect(onQueryChange.mock.calls[1][0].combinator).toBe('or');
   });
 
-  it('should set default value for a rule', async () => {
+  it('sets default value for a rule', async () => {
     const { selectors, onQueryChange } = setup();
     selectors.rerender(
       <QueryBuilder
@@ -1013,9 +1021,9 @@ describe('inputType property in field', () => {
   it('sets the input type', async () => {
     const fields: Field[] = [{ name: 'field1', label: 'Field 1', inputType: 'number' }];
     const onQueryChange = jest.fn<never, [RuleGroupType]>();
-    const { container } = render(<QueryBuilder fields={fields} onQueryChange={onQueryChange} />);
-
-    await user.click(screen.getByTestId(TestID.addRule));
+    const { container } = render(
+      <QueryBuilder fields={fields} onQueryChange={onQueryChange} addRuleToNewGroups />
+    );
 
     expect(container.querySelector('input[type="number"]')).toBeDefined();
   });
@@ -1025,9 +1033,9 @@ describe('valueEditorType property in field', () => {
   it('sets the value editor type', async () => {
     const fields: Field[] = [{ name: 'field1', label: 'Field 1', valueEditorType: 'select' }];
     const onQueryChange = jest.fn<never, [RuleGroupType]>();
-    const { container } = render(<QueryBuilder fields={fields} onQueryChange={onQueryChange} />);
-
-    await user.click(screen.getByTestId(TestID.addRule));
+    const { container } = render(
+      <QueryBuilder fields={fields} onQueryChange={onQueryChange} addRuleToNewGroups />
+    );
 
     expect(container.querySelector(`select.${sc.value}`)).toBeDefined();
   });
@@ -1041,9 +1049,9 @@ describe('operators property in field', () => {
       { name: 'field2', label: 'Field 2', operators },
     ];
     const onQueryChange = jest.fn<never, [RuleGroupType]>();
-    const { container } = render(<QueryBuilder fields={fields} onQueryChange={onQueryChange} />);
-
-    await user.click(screen.getByTestId(TestID.addRule));
+    const { container } = render(
+      <QueryBuilder fields={fields} onQueryChange={onQueryChange} addRuleToNewGroups />
+    );
 
     expect(container.querySelector(`select.${sc.operators}`)).toBeDefined();
     expect(container.querySelectorAll(`select.${sc.operators} option`)).toHaveLength(1);
@@ -1156,6 +1164,69 @@ describe('autoSelectOperator', () => {
   });
 });
 
+describe('valueEditorType "multiselect" default values', () => {
+  const fields: Field[] = [
+    {
+      name: 'field1',
+      label: 'Field 1',
+      valueEditorType: 'multiselect',
+      values: [
+        { name: 'test1', label: 'Test1' },
+        { name: 'test2', label: 'Test2' },
+      ],
+    },
+  ];
+
+  it('does not unnecessarily select a value - string values', async () => {
+    const onQueryChange = jest.fn<never, [RuleGroupType]>();
+    render(<QueryBuilder fields={fields} addRuleToNewGroups onQueryChange={onQueryChange} />);
+    expect(onQueryChange).toHaveBeenCalledWith(
+      expect.objectContaining({
+        id: expect.any(String),
+        combinator: 'and',
+        not: false,
+        rules: [
+          {
+            id: expect.any(String),
+            field: 'field1',
+            operator: '=',
+            value: '',
+            valueSource: 'value',
+          },
+        ],
+      })
+    );
+  });
+
+  it('does not unnecessarily select a value - lists as arrays', async () => {
+    const onQueryChange = jest.fn<never, [RuleGroupType]>();
+    render(
+      <QueryBuilder
+        fields={fields}
+        addRuleToNewGroups
+        onQueryChange={onQueryChange}
+        listsAsArrays
+      />
+    );
+    expect(onQueryChange).toHaveBeenCalledWith(
+      expect.objectContaining({
+        id: expect.any(String),
+        combinator: 'and',
+        not: false,
+        rules: [
+          {
+            id: expect.any(String),
+            field: 'field1',
+            operator: '=',
+            value: [],
+            valueSource: 'value',
+          },
+        ],
+      })
+    );
+  });
+});
+
 describe('addRuleToNewGroups', () => {
   const query: RuleGroupType = { combinator: 'and', rules: [] };
 
@@ -1180,7 +1251,7 @@ describe('addRuleToNewGroups', () => {
 });
 
 describe('showShiftActions', () => {
-  it('should be disabled if rule is locked', async () => {
+  it('is disabled if rule is locked', async () => {
     const onQueryChange = jest.fn<never, [RuleGroupType]>();
     render(
       <QueryBuilder
@@ -1210,7 +1281,7 @@ describe('showShiftActions', () => {
   });
 
   describe('standard rule groups', () => {
-    it('should shift rules', async () => {
+    it('shifts rules', async () => {
       const onQueryChange = jest.fn<never, [RuleGroupType]>();
       render(
         <QueryBuilder
@@ -1237,7 +1308,7 @@ describe('showShiftActions', () => {
       });
     });
 
-    it('should shift rule groups', async () => {
+    it('shifts rule groups', async () => {
       const onQueryChange = jest.fn<never, [RuleGroupType]>();
       render(
         <QueryBuilder
@@ -1270,7 +1341,7 @@ describe('showShiftActions', () => {
   });
 
   describe('independent combinators', () => {
-    it('should shift rulew with independent combinators', async () => {
+    it('shifts rulew with independent combinators', async () => {
       const onQueryChange = jest.fn<never, [RuleGroupTypeIC]>();
       render(
         <QueryBuilder
@@ -1297,7 +1368,7 @@ describe('showShiftActions', () => {
       });
     });
 
-    it('should shift first rule with independent combinators', async () => {
+    it('shifts first rule with independent combinators', async () => {
       const onQueryChange = jest.fn<never, [RuleGroupTypeIC]>();
       render(
         <QueryBuilder
@@ -1326,7 +1397,7 @@ describe('showShiftActions', () => {
 
 describe('showCloneButtons', () => {
   describe('standard rule groups', () => {
-    it('should clone rules', async () => {
+    it('clones rules', async () => {
       const onQueryChange = jest.fn<never, [RuleGroupType]>();
       render(
         <QueryBuilder
@@ -1352,7 +1423,7 @@ describe('showCloneButtons', () => {
       });
     });
 
-    it('should clone rule groups', async () => {
+    it('clones rule groups', async () => {
       const onQueryChange = jest.fn<never, [RuleGroupType]>();
       render(
         <QueryBuilder
@@ -1389,7 +1460,7 @@ describe('showCloneButtons', () => {
   });
 
   describe('independent combinators', () => {
-    it('should clone a single rule with independent combinators', async () => {
+    it('clones a single rule with independent combinators', async () => {
       const onQueryChange = jest.fn<never, [RuleGroupTypeIC]>();
       render(
         <QueryBuilder
@@ -1410,7 +1481,7 @@ describe('showCloneButtons', () => {
       });
     });
 
-    it('should clone first rule with independent combinators', async () => {
+    it('clones first rule with independent combinators', async () => {
       const onQueryChange = jest.fn<never, [RuleGroupTypeIC]>();
       render(
         <QueryBuilder
@@ -1437,7 +1508,7 @@ describe('showCloneButtons', () => {
       });
     });
 
-    it('should clone last rule with independent combinators', async () => {
+    it('clones last rule with independent combinators', async () => {
       const onQueryChange = jest.fn<never, [RuleGroupTypeIC]>();
       render(
         <QueryBuilder
@@ -1495,20 +1566,20 @@ describe('idGenerator', () => {
 });
 
 describe('independent combinators', () => {
-  it('should render a rule group with independent combinators', () => {
+  it('renders a rule group with independent combinators', () => {
     const onQueryChange = jest.fn<never, [RuleGroupTypeIC]>();
     render(<QueryBuilder defaultQuery={{ rules: [] }} onQueryChange={onQueryChange} />);
     expect(screen.getByTestId(TestID.ruleGroup)).toBeDefined();
     expect(onQueryChange.mock.calls[0][0]).not.toHaveProperty('combinator');
   });
 
-  it('should render a rule group with addRuleToNewGroups', async () => {
+  it('renders a rule group with addRuleToNewGroups', async () => {
     render(<QueryBuilder addRuleToNewGroups defaultQuery={{ rules: [] }} />);
     await user.click(screen.getByTestId(TestID.addGroup));
     expect(screen.getByTestId(TestID.rule)).toBeDefined();
   });
 
-  it('should call onQueryChange with query', () => {
+  it('calls onQueryChange with query', () => {
     const onQueryChange = jest.fn<never, [RuleGroupTypeIC]>();
     const dq: RuleGroupTypeIC = { id: 'id', rules: [], not: false };
     render(<QueryBuilder onQueryChange={onQueryChange} defaultQuery={dq} />);
@@ -1516,7 +1587,7 @@ describe('independent combinators', () => {
     expect(onQueryChange.mock.calls[0][0]).toEqual(dq);
   });
 
-  it('should add rules with independent combinators', async () => {
+  it('adds rules with independent combinators', async () => {
     // render(<QueryBuilder defaultQuery={{ rules: [] }} />);
     render(<QueryBuilder defaultQuery={{ rules: [] }} />);
     expect(screen.queryAllByTestId(TestID.combinators)).toHaveLength(0);
@@ -1533,7 +1604,7 @@ describe('independent combinators', () => {
     expect(combinatorSelectors[0]).toHaveValue('or');
   });
 
-  it('should add groups with independent combinators', async () => {
+  it('adds groups with independent combinators', async () => {
     render(<QueryBuilder defaultQuery={{ rules: [] }} />);
     expect(screen.queryAllByTestId(TestID.combinators)).toHaveLength(0);
     await user.click(screen.getByTestId(TestID.addGroup));
@@ -1549,7 +1620,7 @@ describe('independent combinators', () => {
     expect(combinatorSelectors[0]).toHaveValue('or');
   });
 
-  it('should remove rules along with independent combinators', async () => {
+  it('removes rules along with independent combinators', async () => {
     const onQueryChange = jest.fn<never, [RuleGroupTypeIC]>();
     const query: RuleGroupTypeIC = {
       rules: [
@@ -1574,7 +1645,7 @@ describe('independent combinators', () => {
     expect(onQueryChange.mock.calls[2][0].rules[0]).toHaveProperty('value', '3');
   });
 
-  it('should remove groups along with independent combinators', async () => {
+  it('removes groups along with independent combinators', async () => {
     const onQueryChange = jest.fn<never, [RuleGroupTypeIC]>();
     const query: RuleGroupTypeIC = {
       rules: [{ rules: [] }, 'and', { rules: [] }, 'or', { rules: [] }],
@@ -1596,13 +1667,13 @@ describe('independent combinators', () => {
 });
 
 describe('validation', () => {
-  it('should not validate if no validator function is provided', () => {
+  it('does not validate if no validator function is provided', () => {
     render(<QueryBuilder />);
     expect(screen.getByRole('form')).not.toHaveClass(sc.valid);
     expect(screen.getByRole('form')).not.toHaveClass(sc.invalid);
   });
 
-  it('should validate groups if default validator function is provided', async () => {
+  it('validates groups if default validator function is provided', async () => {
     const { container } = render(<QueryBuilder validator={defaultValidator} />);
     await user.click(screen.getByTestId(TestID.addGroup));
     // Expect the root group to be valid (contains the inner group)
@@ -1611,7 +1682,7 @@ describe('validation', () => {
     expect(container.querySelectorAll(`.${sc.ruleGroup}.${sc.invalid}`)).toHaveLength(1);
   });
 
-  it('should use custom validator function returning false', () => {
+  it('uses custom validator function returning false', () => {
     const validator = jest.fn(() => false);
     render(<QueryBuilder validator={validator} />);
     expect(validator).toHaveBeenCalled();
@@ -1619,7 +1690,7 @@ describe('validation', () => {
     expect(screen.getByRole('form')).toHaveClass(sc.invalid);
   });
 
-  it('should use custom validator function returning true', () => {
+  it('uses custom validator function returning true', () => {
     const validator = jest.fn(() => true);
     render(<QueryBuilder validator={validator} />);
     expect(validator).toHaveBeenCalled();
@@ -1627,7 +1698,7 @@ describe('validation', () => {
     expect(screen.getByRole('form')).not.toHaveClass(sc.invalid);
   });
 
-  it('should pass down validationMap to children', () => {
+  it('passes down validationMap to children', () => {
     const valMap: ValidationMap = {
       id: { valid: false, reasons: ['invalid'] },
     };
@@ -1645,12 +1716,12 @@ describe('validation', () => {
 });
 
 describe('disabled', () => {
-  it('should have the correct classname', () => {
+  it('has the correct classname', () => {
     const { container } = render(<QueryBuilder disabled />);
     expect(container.querySelectorAll('div')[0]).toHaveClass(sc.disabled);
   });
 
-  it('should have the correct classname when disabled prop is false but root group is disabled', () => {
+  it('has the correct classname when disabled prop is false but root group is disabled', () => {
     const { container } = render(
       <QueryBuilder query={{ disabled: true, combinator: 'and', rules: [] }} />
     );
@@ -1966,7 +2037,7 @@ describe('value source field', () => {
 });
 
 describe('dynamic classNames', () => {
-  it('should have correct group-based classNames', () => {
+  it('has correct group-based classNames', () => {
     render(
       <QueryBuilder
         fields={[{ name: 'f1', label: 'F1', className: 'custom-fieldBased-class' }]}
@@ -1989,7 +2060,7 @@ describe('dynamic classNames', () => {
   });
 });
 
-describe('dispatchQuery and getQuery', () => {
+describe('redux functions', () => {
   it('gets the query from the store', async () => {
     const onQueryChange = jest.fn<never, [RuleGroupType]>();
     const testFunc = jest.fn();